# Conflicts:
#	src/main/java/com/codecool/marsexploration/logic/ExplorationSimulator.java

+++ src/main/java/com/codecool/marsexploration/data/SimulationInput.java
@@ -2,6 +2,6 @@ package com.codecool.marsexploration.data;
 
 import com.sun.net.httpserver.Authenticator;
 
-public record SimulationInput(String mapPath, Coordinate landing, long timeout, String logPath, SuccessCondition condition) {
+public record SimulationInput(String mapPath, Coordinate landing, int timeout, String logPath, SuccessCondition condition) {
 
 }
